test: Dialect 관련 테스트 추가 및 테스트 패키지 이동
