--- src/test/java/persistence/sql/ddl/generator/meta/NullableTest.java
@@ -1,43 +1,0 @@
-package persistence.sql.ddl.generator.meta;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Id;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
-import persistence.sql.ddl.dialect.H2Dialect;
-
-import java.lang.reflect.Field;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-class NullableTest {
-
-    private static final Class<Person> personClass = Person.class;
-
-
-    // TODO 테스트 추가 수정 필요
-    @ParameterizedTest()
-    @CsvSource({
-            "id,              NOT NULL",
-            "notNullCol,      NOT NULL",
-            "noAnnotationCol, NULL",
-            "nullableCol,     NULL",
-    })
-    void nullableSQL(String fieldName, String sql) throws NoSuchFieldException {
-        Field field = personClass.getDeclaredField(fieldName);
-        assertThat(new H2Dialect().getColumnNullableType(field)).isEqualTo(sql);
-    }
-
-    private static class Person {
-        @Id
-        private Long id;
-
-        @Column(nullable = false)
-        private String notNullCol;
-
-        @Column(nullable = true)
-        private String nullableCol;
-
-        private String noAnnotationCol;
-    }
-}
